pyproject.toml
[project]
name = "apple-clip"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub[hf-xet]>=0.36.0",
    "matplotlib>=3.10.8",
    "onnxruntime-gpu>=1.23.2",
    "onnxscript>=0.5.7",
    "open-clip-torch>=3.2.0",
    "optimum[onnxruntime]>=2.1.0",
    "pillow>=12.1.0",
    "testcontainers>=4.14.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers==4.57.6",
]

[[tool.uv.index]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-gpu" }
torchvision = { index = "pytorch-gpu" }
